requirements.txt
praw
dateparser
tqdm
pyyaml